+++ backend/agents/00-index/RULES.md
@@ -14,6 +14,8 @@
 
 - 현재 코드에서 쓰이는 패턴과 비슷하게 만드세요.
 
+- 인증이나 사용자를 정보가 필요한 로직은 JWT 토큰을 이용해 인증하도록 하세요.
+
 # 코딩 로그
 1. 진행중인 코딩을 /backend/agents/60-tasks폴더 안에 새로운 task에 대한 md파일을 만들고 기록하세요.
