+++ backend/models/__init__.py
@@ -1,6 +1,16 @@
 """ORM 모델 패키지."""
 
-from .entities import Membership, Organization, Workspace, WorkspaceType, User
+from .entities import (
+    Membership,
+    Organization,
+    Workspace,
+    WorkspaceType,
+    User,
+    NotionOauthCredentials,
+    DataSource
+)
+
+
 
 __all__ = [
     "Membership",
@@ -8,4 +18,6 @@ __all__ = [
     "Workspace",
     "WorkspaceType",
     "User",
+    "NotionOauthCredentials",
+    "DataSource",
 ]
