+++ backend/models/entities.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from datetime import datetime
 from enum import Enum
 
-from sqlalchemy import BigInteger, Column, DateTime, String, Boolean
+from sqlalchemy import BigInteger, Column, DateTime, String, Boolean, Text, JSON
 
 from utils.db import Base
 
@@ -63,4 +63,33 @@ class Membership(Base):
     organization_idx = Column(BigInteger, nullable=False)
     user_idx = Column(BigInteger, nullable=False)
     role = Column(String(50), nullable=False, default="member")
+    created = Column(DateTime, nullable=False, default=datetime.utcnow)
+
+class NotionOauthCredentials(Base):
+    __tablename__ = "notion_oauth_credentials"
+
+    idx                   = Column(BigInteger, primary_key=True, autoincrement=True)
+    user_idx              = Column(BigInteger, nullable=False)         # FK: users.idx
+    data_source_idx       = Column(BigInteger, nullable=False)         # FK: data_sources.idx
+    provider              = Column(String(50), nullable=False, default="notion")
+    bot_id                = Column(String(100), nullable=False)
+    provider_workspace_id = Column(String(64), nullable=True)          # Notion workspace_id
+    workspace_name        = Column(String(255), nullable=True)
+    workspace_icon        = Column(String(1000), nullable=True)
+    token_type            = Column(String(20), nullable=False, default="bearer")
+    access_token          = Column(Text, nullable=False)
+    refresh_token         = Column(Text, nullable=True)
+    expires               = Column(DateTime, nullable=True)            # MySQL DATETIME (naive)
+    created               = Column(DateTime, nullable=False, default=lambda: datetime.utcnow())
+    updated               = Column(DateTime, nullable=False, default=lambda: datetime.utcnow())
+    provider_payload      = Column(JSON, nullable=True)
+
+class DataSource(Base):
+    __tablename__ = "data_sources"
+    idx = Column(BigInteger, primary_key=True, autoincrement=True)
+    workspace_idx = Column(BigInteger, nullable=False)  # FK: workspaces.idx
+    type = Column(String(20), nullable=False)           # 'notion'
+    name = Column(String(200), nullable=False)
+    status = Column(String(20), nullable=False, default="connected")  # connected/disconnected/error
+    synced = Column(DateTime, nullable=True)
     created = Column(DateTime, nullable=False, default=datetime.utcnow)
